5.12. Модуляция https://up.htmlacademy.ru/javascript/23/tasks/9
